Split Vercel deployment step
Break the Vercel deployment implementation into focused partial files for orchestration, plan generation, CLI handling, state, Build Output/VCR setup, environment mapping, model validation, and shared types.
